+++ demo/headings.ts
@@ -26,6 +26,7 @@ const headings: HeadingDefinition[] = [
   {
     level: 1,
     title: 'Overview',
+    color: '#0d0',
   },
   {
     level: 2,
@@ -42,11 +43,11 @@ const headings: HeadingDefinition[] = [
     level: 2,
     title: 'Features',
     url: '/features',
-    topNavigation: '🌱 Get Started',
   },
   {
     level: 1,
     title: 'Another Section',
+    color: '#d00',
   },
   {
     level: 2,
@@ -80,6 +81,7 @@ const headings: HeadingDefinition[] = [
   {
     level: 1,
     title: 'API',
+    color: '#00d',
     menuModalFullWidth: true,
   },
   {