+++ demo/pages/index/+Page.tsx
@@ -28,7 +28,11 @@ function Page() {
 function Header() {
   return (
     <>
-      <h1 style={{ textAlign: 'center', fontSize: '3.4em' }}>Next Generation<br/>Frontend Framework</h1>
+      <h1 style={{ textAlign: 'center', fontSize: '3.4em' }}>
+        Next Generation
+        <br />
+        Frontend Framework
+      </h1>
     </>
   )
 }