+++ src/utils/PassTrough.tsx
@@ -1,0 +1,7 @@
+export { PassThrough }
+
+import React from 'react'
+
+function PassThrough({ children }: any) {
+  return <>{children}</>
+}